--HG--
branch : akd

+++ scripts/pymod-process.py
@@ -1256,7 +1256,7 @@ def generate_getset_code(tp):
 	getter = """
 local PyObject * %(getname)s(PyObject *obj, void *v)
 {
-	int offset = (int)v;
+	long offset = (long)v;
 	%(ctype)s *ptr = (%(ctype)s*)(((char*)obj)+offset);
 """ % vars()
 	try:
@@ -1273,7 +1273,7 @@ local PyObject * %(getname)s(PyObject *obj, void *v)
 	setter = """
 local int %(setname)s(PyObject *obj, PyObject *newval, void *v)
 {
-	int offset = (int)v;
+	long offset = (long)v;
 	%(ctype)s *ptr = (%(ctype)s*)(((char*)obj)+offset);
 """ % vars()
 
@@ -1281,7 +1281,7 @@ local int %(setname)s(PyObject *obj, PyObject *newval, void *v)
 		setter = """
 local int %(setname)s(PyObject *obj, PyObject *newval, void *v)
 {
-	int offset = (int)v;
+	long offset = (long)v;
 	int *ptr = (int*)(((char*)obj)+offset);
 """ % vars()
