+++ src/pymod/pymod.c
@@ -195,7 +195,7 @@ local PyObject * cvt_c2p_banner(Banner *b)
 local int cvt_p2c_banner(PyObject *o, Banner **bp)
 {
 	char *buf;
-	int len = -1;
+	Py_ssize_t len = -1;
 	PyString_AsStringAndSize(o, &buf, &len);
 	if (len == sizeof(Banner))
 	{
@@ -835,8 +835,8 @@ local void py_newplayer(Player *p, int isnew)
 	else
 	{
 		if (d->obj->ob_refcnt != 1)
-			lm->Log(L_ERROR, "<pymod> there are %d remaining references to a player object!",
-					d->obj->ob_refcnt);
+			lm->Log(L_ERROR, "<pymod> there are %lu remaining references to a player object!",
+					(unsigned long)d->obj->ob_refcnt);
 
 		/* this stuff would usually be done in dealloc, but I want to
 		 * make sure player objects for players who are gone are
@@ -869,8 +869,8 @@ local void py_aaction(Arena *a, int action)
 	if (action == AA_POSTDESTROY && d->obj)
 	{
 		if (d->obj->ob_refcnt != 1)
-			lm->Log(L_ERROR, "<pymod> there are %d remaining references to an arena object!",
-					d->obj->ob_refcnt);
+			lm->Log(L_ERROR, "<pymod> there are %lu remaining references to an arena object!",
+					(unsigned long)d->obj->ob_refcnt);
 
 		/* see notes for py_newplayer as to why this is done here. */
 		d->obj->a = NULL;
@@ -1353,7 +1353,7 @@ local int get_player_data(Player *p, void *data, int len, void *v)
 	struct pypersist_ppd *pyppd = v;
 	PyObject *val, *pkl;
 	const void *pkldata;
-	int pkllen;
+	Py_ssize_t pkllen;
 
 	val = PyObject_CallMethod(pyppd->funcs, "get", "(O&)", cvt_c2p_player, p);
 	if (!val)
@@ -1386,8 +1386,8 @@ local int get_player_data(Player *p, void *data, int len, void *v)
 	{
 		Py_DECREF(pkl);
 		lm->Log(L_WARN, "<pymod> persistent data getter returned more "
-				"than %d bytes of data (%d allowed)",
-				pkllen, len);
+				"than %lu bytes of data (%d allowed)",
+				(unsigned long)pkllen, len);
 		return 0;
 	}
 
@@ -1479,7 +1479,7 @@ local int get_arena_data(Arena *a, void *data, int len, void *v)
 	struct pypersist_apd *pyapd = v;
 	PyObject *val, *pkl;
 	const void *pkldata;
-	int pkllen;
+	Py_ssize_t pkllen;
 
 	val = PyObject_CallMethod(pyapd->funcs, "get", "(O&)", cvt_c2p_arena, a);
 	if (!val)
@@ -1512,8 +1512,8 @@ local int get_arena_data(Arena *a, void *data, int len, void *v)
 	{
 		Py_DECREF(pkl);
 		lm->Log(L_WARN, "<pymod> persistent data getter returned more "
-				"than %d bytes of data (%d allowed)",
-				pkllen, len);
+				"than %lu bytes of data (%d allowed)",
+				(unsigned long)pkllen, len);
 		return 0;
 	}
 
@@ -1794,8 +1794,8 @@ local int unload_py_module(mod_args_t *args)
 
 	if (mod->ob_refcnt != 2)
 	{
-		lm->Log(L_WARN, "<pymod> there are %d remaining references to module %s",
-				mod->ob_refcnt, mname);
+		lm->Log(L_WARN, "<pymod> there are %lu remaining references to module %s",
+				(unsigned long)mod->ob_refcnt, mname);
 		return MM_FAIL;
 	}
